requirements.txt
fire
triton
pydantic==2.8.2
accelerate==0.27.2
deepspeed==0.8.1
huggingface-hub==0.23.4
imageio==2.25.1
numpy==1.21.6
omegaconf==2.3.0
pandas==1.3.5
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
transformers==4.42.4
xformers==0.0.16
safetensors
diffusers==0.28.2
sentencepiece==0.1.99
gradio==4.37.2
opencv-python
einops
timm
onnxruntime
wandb
bitsandbytes
insightface